+++ dbdaora/hash/service/__init__.py
@@ -1,12 +1,9 @@
 from typing import Any
 
-from google.cloud.datastore import Key as DatastoreKey
-
-from ..data_sources.fallback.mongodb import Key as MongoKey
-from ..entity import Entity
-from ..keys import FallbackKey
-from ..service import Service
-from .repositories import HashData
+from ...entity import Entity
+from ...keys import FallbackKey
+from ...service import Service
+from ..repositories import HashData
 
 
 class HashService(Service[Entity, HashData, FallbackKey]):
@@ -37,11 +34,3 @@ class HashService(Service[Entity, HashData, FallbackKey]):
                     return None
 
         return entity
-
-
-class DatastoreHashService(HashService[Entity, DatastoreKey]):
-    ...
-
-
-class MongoHashService(HashService[Entity, MongoKey]):
-    ...