+++ dbdaora/hash/service/datastore.py
@@ -1,0 +1,8 @@
+from google.cloud.datastore import Key as DatastoreKey
+
+from ...entity import Entity
+from . import HashService
+
+
+class DatastoreHashService(HashService[Entity, DatastoreKey]):
+    ...