+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix service
+
 ### 0.17.1 - 2020-07-21
 
  - Fix service