+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ### UNRELEASED
 
+ - Fix Hash services conditional imports
+
  - Fix Service class to be generic
 
 ### 0.17.2 - 2020-07-21